move config to root from core
